fix: consolidate dependent initialization
The issue here was that we used dependents both for secondary resources
for actual halkyon resources but also at a stage where they were not
associated to a halkyon resource during the controller initialization
to register watches. Dependents are now properly initialized in
FetchAndCreateNew and we pass a callback that needs to be called when
the dependent needs to be watched. BaseResource provides a solid default
implementation.